requirements.txt
streamlit
streamlit_js
openai
python-dotenv
requests